requirements.txt
astroid==2.2.0
autopep8==1.4.3
boto3==1.9.106
botocore==1.12.106
colorama==0.4.1
Django==2.1.7
django-crispy-forms==1.7.2
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
isort==4.3.9
jmespath==0.9.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pycodestyle==2.5.0
pylint==2.3.0
python-dateutil==2.8.0
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
